Changing script to smoke test and creating basic folder structure
